refactor: clean-up code
